move JavaNameShortener to :generator
extract JavaNames.getRawClassName to keep in :bnf-language
