+++ bnf-language/src/org/intellij/grammar/analysis/BnfFirstNextAnalyzer.java
@@ -18,7 +18,6 @@ import com.intellij.util.containers.ContainerUtil;
 import com.intellij.util.containers.JBIterable;
 import it.unimi.dsi.fastutil.objects.ObjectOpenCustomHashSet;
 import org.intellij.grammar.KnownAttribute;
-import org.intellij.grammar.generator.ParserGeneratorUtil;
 import org.intellij.grammar.generator.RuleGraphHelper;
 import org.intellij.grammar.psi.*;
 import org.intellij.grammar.psi.impl.GrammarUtil;
