+++ bnf-language/src/org/intellij/grammar/generator/RuleGraphHelper.java
@@ -28,10 +28,12 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
-import static org.intellij.grammar.psi.BnfAttributes.*;
-import static org.intellij.grammar.psi.BnfAst.*;
-import static org.intellij.grammar.generator.ParserGeneratorUtil.*;
+import static org.intellij.grammar.generator.ParserGeneratorUtil.getChildExpressions;
+import static org.intellij.grammar.generator.ParserGeneratorUtil.topoSort;
 import static org.intellij.grammar.generator.RuleGraphHelper.Cardinality.*;
+import static org.intellij.grammar.psi.BnfAst.getEffectiveType;
+import static org.intellij.grammar.psi.BnfAttributes.getAttribute;
+import static org.intellij.grammar.psi.BnfAttributes.getRootAttribute;
 import static org.intellij.grammar.psi.impl.GrammarUtil.collectMetaParameters;
 import static org.intellij.grammar.psi.impl.GrammarUtil.isDoubleAngles;
