+++ bnf-language/src/org/intellij/grammar/psi/impl/GrammarPsiImplUtil.java
@@ -21,9 +21,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import static org.intellij.grammar.psi.BnfAst.*;
-import static org.intellij.grammar.generator.ParserGeneratorUtil.*;
-import static org.intellij.grammar.psi.BnfAttributes.*;
+import static org.intellij.grammar.generator.ParserGeneratorUtil.findRuleImplMethods;
+import static org.intellij.grammar.psi.BnfAttributes.getAttribute;
+import static org.intellij.grammar.psi.BnfAttributes.getRootAttribute;
 
 /**
  * @author gregsh
