+++ docs/architecture.md
@@ -21,9 +21,9 @@ Notable shape: `:bnf-language` and `:jflex-language` are **peers** — neither d
 |-------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | `:base`           | Shared infrastructure: `GrammarKitBundle` (i18n), `BnfIcons`, message resources.                                                                                                                                                                                                                 |
 | `:parser-runtime` | Language-agnostic parser runtime: `GeneratedParserUtilBase` plus `config.Options`/`config.Option`. No BNF or JFlex specifics.                                                                                                                                                                    |
-| `:bnf-language`   | BNF language model: `psi/`, `parser/`, `analysis/`, `java/`, language registration (`BnfLanguage`, `BnfFileType`, …), `KnownAttribute`, `LightPsi`. Plus PSI-shared rendering and naming primitives (`Renderer`, `JavaRenderer`, `KotlinRenderer`, `Case`, `NameFormat`, `RuleGraphHelper`, `ParserGeneratorUtil`, `JavaNameShortener`, `*BnfConstants`) that `:generator` and PSI both consume. Owns `grammars/Grammar.bnf` and BNF `gen/`. |
+| `:bnf-language`   | BNF language model: `psi/`, `parser/`, `analysis/`, `java/`, language registration (`BnfLanguage`, `BnfFileType`, …), `KnownAttribute`, `LightPsi`. Plus PSI-shared rendering and naming primitives (`Renderer`, `JavaRenderer`, `KotlinRenderer`, `Case`, `NameFormat`, `NameShortener`, `JavaNames`, `RuleGraphHelper`, `ParserGeneratorUtil`, `*BnfConstants`) that `:generator` and PSI both consume. Owns `grammars/Grammar.bnf` and BNF `gen/`. |
 | `:jflex-language` | JFlex language support: PSI, parser, editor, file type. True peer of `:bnf-language` — depends only on `:base` and `:parser-runtime`. Owns `grammars/JFlex.bnf` and JFlex `gen/`.                                                                                                                  |
-| `:generator`      | Pure code emission: `Generator`, `JavaParserGenerator`, `KotlinParserGenerator`, `ExpressionGeneratorHelper`, `ExpressionHelper`, `ExpressionInfo`, `RuleMethodsHelper`, `RuleInfo`, `OperatorInfo`, `OperatorType`, `NodeCalls`, `Names`, `FilePrinter`, `OutputOpener`, `GenOptions`, `KotlinNameShortener`, `KotlinPlatformConstants`, plus `Main.java` (headless CLI). Depends on `:bnf-language`. |
+| `:generator`      | Pure code emission: `Generator`, `JavaParserGenerator`, `KotlinParserGenerator`, `ExpressionGeneratorHelper`, `ExpressionHelper`, `ExpressionInfo`, `RuleMethodsHelper`, `RuleInfo`, `OperatorInfo`, `OperatorType`, `NodeCalls`, `Names`, `FilePrinter`, `OutputOpener`, `GenOptions`, `java/JavaNameShortener`, `kotlin/{KotlinNameShortener, KotlinPlatformConstants}`, plus `Main.java` (headless CLI). Depends on `:bnf-language`. |
 | Root project      | The IDE plugin itself: `actions/`, `editor/`, `inspection/`, `intention/`, `refactor/`, `search/`, `livePreview/`, `diagram/`, top-level UI providers, `plugin.xml`, signing, publishing.                                                                                                          |
 
 The IntelliJ Platform Gradle plugin (v2) auto-detects subproject dependencies and bundles each module's jar into the plugin distribution.
@@ -57,11 +57,11 @@ In `:bnf-language`:
 - `psi/` — PSI interfaces/impl plus the extracted BNF-semantic utilities `BnfRules`, `BnfAttributes`, `BnfAst`.
 - `analysis/` — `BnfFirstNextAnalyzer` computes FIRST/FOLLOW sets used by precedence handling, recovery, and inspections.
 - `java/` — `JavaHelper`: ASM-based Java-class introspection for target type resolution.
-- `generator/` (split package, partial) — `Renderer` family, `Case`, `NameFormat`, `RuleGraphHelper`, `ParserGeneratorUtil`, `JavaNameShortener`, `*BnfConstants`. These are PSI-shared naming/semantic primitives that happen to live in the generator package historically.
+- `generator/` (split package, partial) — `Renderer` family, `Case`, `NameFormat`, `NameShortener`, `RuleGraphHelper`, `ParserGeneratorUtil`, `*BnfConstants`, plus `generator/java/{JavaBnfConstants, JavaNames, JavaRenderer}` and `generator/kotlin/{KotlinBnfConstants, KotlinRenderer}`. These are PSI-shared naming/semantic primitives that happen to live in the generator package historically.
 
 In `:generator`:
 
-- `generator/` — Code emission (`Generator`, `Java/KotlinParserGenerator`, `ExpressionGeneratorHelper`, `RuleMethodsHelper`, etc.), output sinks (`FilePrinter`, `OutputOpener`), generator config (`GenOptions`, `Names`, `KotlinPlatformConstants`).
+- `generator/` — Code emission (`Generator`, `Java/KotlinParserGenerator`, `ExpressionGeneratorHelper`, `RuleMethodsHelper`, etc.), output sinks (`FilePrinter`, `OutputOpener`), generator config (`GenOptions`, `Names`), plus `generator/kotlin/{KotlinNameShortener, KotlinPlatformConstants}`.
 - `Main.java` — Headless CLI entry point.
 
 In root project (IDE plugin):
